p4

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const logger = require('../utils/logger')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
   })
 
@@ -17,6 +19,9 @@ blogsRouter.get('/:id', async (request, response) => {
   
 blogsRouter.post('/', async (request, response) => {
     const blog = new Blog(request.body)
+    const user = request.user
+    
+    blog.user = user
 
     if (!blog.title) {
         return response.status(400).json({
@@ -27,14 +32,26 @@ blogsRouter.post('/', async (request, response) => {
             error: 'url missing'
         })
     }
-    
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    
     response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    logger.info('blog user', blog.user.toString())
+    logger.info('user', user._id.toString())
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }
+    return response.status(401).json({
+      error: 'unauthorize to delete'
+    })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
